New changes: re-ordering new functions such as TestMaild and TestAnalysisd Changing wm_control_main in order to increase its capabilities
