+++ src/config/config.c
@@ -13,6 +13,7 @@
 #include "shared.h"
 #include "os_xml/os_xml.h"
 #include "config.h"
+#include "global-config.h"
 
 /* Prototypes */
 static int read_main_elements(const OS_XML *xml, int modules,
@@ -386,4 +387,39 @@ int ReadConfig(int modules, const char *cfgfile, void *d1, void *d2)
     OS_ClearNode(node);
     OS_ClearXML(&xml);
     return (0);
+}
+
+int Test_Analysisd(const char * path) {
+    int fail = 0;
+    _Config * test_config;
+    os_calloc(1, sizeof(_Config), test_config);
+
+    int modules = CGLOBAL | CRULES | CALERTS | CCLUSTER;
+    if( (ReadConfig(modules, path, test_config, NULL) < 0) || (ReadConfig(CLABELS, path, &(test_config->labels), NULL) < 0) ) {
+        merror(RCONFIG_ERROR,"Analysisd", path);
+		fail = 1;
+    }
+
+    if(!fail) {
+        test_config->min_rotate_interval = getDefine_Int("analysisd", "min_rotate_interval", 10, 86400);
+
+        if (test_config->rotate_interval && (test_config->rotate_interval < test_config->min_rotate_interval || test_config->rotate_interval > 86400)) {
+            merror("Rotate interval setting must be between %d seconds and one day.", test_config->min_rotate_interval);
+            fail = 1;
+        }
+
+        if (test_config->max_output_size && (test_config->max_output_size < 1000000 || test_config->max_output_size > 1099511627776)) {
+            merror("Maximum output size must be between 1 MiB and 1 TiB.");
+            fail = 1;
+        }
+    }
+
+    /* Free memory */
+    config_free(test_config);
+
+    if(fail) {
+        return -1;
+    }
+
+    return 0;
 }
