+++ src/os_execd/execd.c
@@ -735,12 +735,4 @@ error:
     return ret_val;
 }
 
-int Test_Execd(const char * path) {
-    if(ExecdConfig(path) < 0) {
-        merror_exit(CONFIG_ERROR, path);
-    }
-
-    return 0;
-}
-
 #endif /* !WIN32 */
